added mixed precision directory
