+++ examples/offline_inference_tpu.py
@@ -19,9 +19,8 @@ sampling_params = SamplingParams(temperature=0.7,
 
 # Set `enforce_eager=True` to avoid ahead-of-time compilation.
 # In real workloads, `enforace_eager` should be `False`.
-# llm = LLM(model="neuralmagic/Meta-Llama-3.1-8B-Instruct-quantized.w8a8",
-llm = LLM(model="nm-testing/Meta-Llama-3-8B-Instruct-W8-Channel-A8-Dynamic-Asym-Per-Token-Test",
-          enforce_eager=False,
+llm = LLM(model="neuralmagic/Meta-Llama-3.1-8B-Instruct-quantized.w8a8",
+          enforce_eager=True,
           max_model_len=1024)
 outputs = llm.generate(prompts, sampling_params)
 for output, answer in zip(outputs, answers):
